+++ .idea/misc.xml
@@ -12,6 +12,7 @@
         <entry key="../../../../layout/compose-model-1662878813795.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1662879158655.xml" value="0.1363247863247863" />
         <entry key="../../../../layout/compose-model-1662879185010.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1662880994572.xml" value="0.2568376068376068" />
         <entry key="..\:/Users/user/Documents/GitHub/Took-Android/presentation/src/main/res/drawable/ic_checked.xml" value="0.228" />
         <entry key="presentation/src/main/res/drawable/ic_bg_blur.xml" value="0.1195" />
         <entry key="presentation/src/main/res/layout/nfc_tech_filter.xml" value="0.1933876811594203" />
